kernel: pci: Add PCI lookup routine
This commit adds a new 'lookup_type_t' for representing the various
kinds of lookups one can perform on the PCI bus. The actual code to
drive the lookups is exposed as the pci_bus_lookup() routine.

When doing a lookup of a PCI device, you must specify the kind of lookup
to perform. If a PCI_LU_CLASSREV is requested, only the PCI class and
subclass fields in the PCI lookup are used.

If the PCI_LU_VENDEV is requested, only the PCI vendor ID and device ID
are used in the lookup.

Signed-off-by: Ian Moffett <ian@osmora.org>