kern: pci: Implement PCI leaf drivers + AHCI stub
This commit introduces the support for having PCI specific device
drivers while providing a tiny AHCI stub that logs the existence of the
host bus adapter. The PCI bus itself has a driver for enumerating the
devices and associating a specific driver to it, the device drivers
themselve must advertise themselves to the PCI driver.

Signed-off-by: Ian Moffett <ian@osmora.org>