kern: ahci: Ensure HBA is in AHCI-mode after reset
The global host control register (described by section 3.1.2 in the AHCI
spec) has an HBA reset bit which resets *everything* to a known state.
This means that after the reset, we'll have to indicate to the HBA that
we want to speak to it over the AHCI interface. We do it once before the
reset if needed, then we do it again after.

Signed-off-by: Ian Moffett <ian@osmora.org>