kernel/amd64: cpu: Implement 'ioapic_common' stub
This implementation walks through the entire table of registered
interrupt handlers and calls each driver handler.

The driver is to return a value of 1 if the interrupt is theirs and has
been handled, otherwise it is to return 0. The handling core is to
respoond to a 1 by exiting the table walk early, and responds to a 0 by
continuing.

Signed-off-by: Ian Moffett <ian@osmora.org>